+++ Zara/ejbModule/server/entidades/OfAD/ItemOfAD.java
@@ -25,8 +25,8 @@ public class ItemOfAD implements Serializable{
 	
 	private int id;
 	private Articulo articulo;
-	private double precioLista;
-	private double precioOferta;
+	private float precioLista;
+	private float precioOferta;
 	private OfAD ofad;
 	
 	public ItemOfAD() {
@@ -42,19 +42,19 @@ public class ItemOfAD implements Serializable{
 		this.articulo = articulo;
 	}
 	
-	public double getPrecioLista() {
+	public float getPrecioLista() {
 		return precioLista;
 	}
 	
-	public void setPrecioLista(double precioLista) {
+	public void setPrecioLista(float precioLista) {
 		this.precioLista = precioLista;
 	}
 	
-	public double getPrecioOferta() {
+	public float getPrecioOferta() {
 		return precioOferta;
 	}
 	
-	public void setPrecioOferta(double precioOferta) {
+	public void setPrecioOferta(float precioOferta) {
 		this.precioOferta = precioOferta;
 	}
